requirements.txt
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.1
pillow==10.3.0
boto3==1.34.81
Unidecode==1.3.8
git+https://github.com/YaserFazil/dynamorm.git
streamlit==1.33.0
marshmallow==3.21.1
streamlit-local-storage==0.0.23